stub: merge header into stub_gssapi.c
