+++ lib/curl_gssapi.c
@@ -147,7 +147,7 @@ void Curl_gss_log_error(struct Curl_easy *data, const char *prefix,
 
   display_gss_error(minor, GSS_C_MECH_CODE, buf, len);
 
-  infof(data, "%s%s", prefix, buf);
+  infof(data, "%s%s (%d, %d)", prefix, buf, major, minor);
 #ifdef CURL_DISABLE_VERBOSE_STRINGS
   (void)data;
   (void)prefix;
